drop precision in shaders remove messaging py<->js report user shader folder simple noise_hash for seed support shader params extended more shader controls for exposing seed noise

+++ glsl/.lib/_.vert
@@ -2,7 +2,8 @@
 // desc: draws 2 triangles as a quad for a surface to manipulate
 
 #version 460
-precision highp float;
+
+precision mediump float;
 
 void main()
 {
